[REFACTOR] 리스트 뷰 파일 구조 변경
