+++ js/main/listView/listView.js
@@ -2,38 +2,15 @@ import {
   makeCategory,
   makeRandomNews,
   transformMainNews,
-} from "./handle/handleData.js";
-import { addEventToBtn } from "./handle/handleBtn.js";
+} from "./feature/handleNewsData.js";
+import { addEventToBtn } from "./feature/handleBtnEvent.js";
 import { fetchData } from "../../utils/fetchData.js";
-
-let category = [];
-let news_by_category = {};
+import { transformNews } from "./feature/transformNews.js";
 
 async function initListView() {
   const news = await fetchData(".././assets/news.json");
-  makeNewsByCategory(news);
+  transformNews(news);
   makeListView();
-}
-
-function makeNewsByCategory(news) {
-  news.forEach((item) => {
-    if (!category.includes(item.category)) category.push(item.category);
-  });
-  category.forEach((item) => {
-    news_by_category[item] = [];
-  });
-  news.forEach((item) => {
-    news_by_category[item.category].push({
-      name: item.name,
-      src: item.lightSrc,
-      isSub: item.isSub,
-      editDate: item.editDate,
-      thumbSrc: item.thumbSrc,
-      headTitle: item.main_title,
-      subTitle: item.subTitle,
-      copyRight: `${item.name} 언론사에서 직접 편집한 뉴스입니다.`,
-    });
-  });
 }
 
 function makeListView() {
@@ -43,4 +20,4 @@ function makeListView() {
   addEventToBtn();
 }
 
-export { initListView, category, news_by_category };
+export { initListView };
